STEP #6 COMPLETE - Working end-to-end ML pipeline for triathlete race time prediction
 MAJOR MILESTONE: Complete ML pipeline successfully trains and saves models

Core Implementation (Steps 1-6):
- Data extraction: Loads 62,063 race results from PostgreSQL
- Feature engineering: 46 features with rolling averages, DNF handling, event parsing
- Label generation: Creates next_time_sec targets for 50,738 samples
- Preprocessing: ColumnTransformer with StandardScaler and PCA
- Training: LinearRegression and RandomForest with GroupKFold CV
- Model persistence: Saves trained models and JSON results

Bug Fixes Applied:
- Fixed pandas Index object compatibility (column name string conversion)
- Fixed boolean dtype compatibility (convert one-hot encoded features to int)
- Fixed JSON serialization (convert numpy types to Python native types)
- Fixed argument parser formatting and syntax errors

Initial Results (Sprint Distance):
- 32,903 training samples processed successfully
- Linear Regression: 1401.54  16.06 seconds MAE
- Random Forest: 1425.58  16.53 seconds MAE
- Baseline: 1292.14 seconds MAE
- Models saved to models/ directory with complete results tracking

Next: Model performance optimization to reach 600 sec MAE target

+++ train.py
@@ -171,18 +171,17 @@ class TriathlonMLTrainer:
                     X_processed,
                     y,
                     cv=cv,
-                    groups=groups,
-                    scoring=neg_mae_scorer,
+                    groups=groups,                    scoring=neg_mae_scorer,
                     n_jobs=-1,
                 )
 
                 mae_scores = -cv_scores
 
                 results[model_name] = {
-                    "mae_mean": mae_scores.mean(),
-                    "mae_std": mae_scores.std(),
+                    "mae_mean": float(mae_scores.mean()),
+                    "mae_std": float(mae_scores.std()),
                     "mae_scores": mae_scores.tolist(),
-                    "n_samples": len(df_dist),
+                    "n_samples": int(len(df_dist)),
                 }
 
                 logger.info(
@@ -234,9 +233,7 @@ class TriathlonMLTrainer:
 
         # Athlete-specific mean baseline
         athlete_means = df_dist.groupby("athlete_id")["TotalTime_sec"].mean()
-        baseline_preds = df_dist["athlete_id"].map(athlete_means)
-
-        # Calculate MAE where both prediction and target are available
+        baseline_preds = df_dist["athlete_id"].map(athlete_means)        # Calculate MAE where both prediction and target are available
         valid_mask = ~df_dist["next_time_sec"].isna() & ~baseline_preds.isna()
 
         if valid_mask.sum() > 0:
@@ -244,7 +241,7 @@ class TriathlonMLTrainer:
                 df_dist.loc[valid_mask, "next_time_sec"]
                 - baseline_preds.loc[valid_mask]
             ).mean()
-            return {"athlete_mean_baseline_mae": mae, "n_samples": valid_mask.sum()}
+            return {"athlete_mean_baseline_mae": float(mae), "n_samples": int(valid_mask.sum())}
 
         return {}
